backend/general-analysis-service/requirements.txt
# Dependencies for the General Analysis Service (v5.2.0)
functions-framework==3.9.2
firebase-admin==7.0.0
google-generativeai==0.8.5
feedparser==6.0.11
requests==2.32.4
pytz==2024.1